+++ src/views/common/breadcrumbs.blade.php
@@ -1,9 +1,0 @@
-            <ul class="breadcrumb">
-            @foreach ($breadcrumbs as $uri => $label)
-                @if ( $uri === 'active' )
-                <li class="active">{{ $label }}</a></li>
-                @else
-                <li><a href="{{ URL::to('admin/'.$uri) }}">{{ $label }}</a> <span class="divider">/</span></li>
-                @endif
-            @endforeach
-            </ul>