+++ src/views/theme/common/foot.blade.php
@@ -1,0 +1,4 @@
+        <script src="http://code.jquery.com/jquery.js"></script>
+        <script src="{{ URL::to($appScript) }}"></script>
+    </body>
+</html>