+++ src/views/theme/common/head.blade.php
@@ -1,0 +1,8 @@
+<!doctype html>
+<html>
+    <head>
+        <title>@yield('title')</title>
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link href="{{ URL::to($appStyle) }}" rel="stylesheet" />
+    </head>
+    <body>