+++ src/views/theme/home/login.blade.php
@@ -1,4 +1,4 @@
-@extends('mothership::layouts.blank')
+@extends('mothership::theme.layouts.blank')
 
 @section('title')
     Please Login